Move AuthenticationTypes and RemoteAuthentication from SharpPcap.LibPcap to SharpPcap
